Issue 7: Added food resource, consumption, and collection to the game.
The values aren't balance yet, and being at zero energy doesn't do anything yet. As part of this, added lower text to locations, and an antUpdate which happens
much less frequently than update. To transport the food a general way of carrying and giving things was added.

Ants try to eat food when they enter a nest, exit a nest, and use the stored food in the nest instead of their energy when in a nest.