Issue 6: Ant death added
General Changes:
Changed how ants are tracked among locations, so that ants are properly removed from assignments when they die.
Ants have a max and min life in seconds, with a uniform distribution over that range.
If an ant is starving, it ages twice as quickly, this is just a start, and likely will change.
Changed ant initialization to allow for a pool of ant objects that are used for living ants, and to which they return when they die